Report first lineno for decorated objects on Python3.8+
