+++ vulture/lines.py
@@ -1,4 +1,5 @@
 import ast
+import sys
 
 
 def _get_last_child_with_lineno(node):
@@ -63,3 +64,19 @@ def get_last_line_number(node):
             except AttributeError:
                 pass
         node = last_child
+
+
+def get_first_line_number(first_node, typ):
+    decorated_typs = ("class", "function", "property")
+    if typ in decorated_typs and sys.version_info >= (3, 8):
+        # From Python3.8 onwards, lineno for decorated objects is the
+        # line at which the object definition starts, which is
+        # different from what Python < 3.8 did -- reporting the lineno
+        # of the first decorator. To preserve this behaviour of Vulture
+        # for newer Python versions, which is also more accurate for
+        # counting the size of the unused code chunk (if the property
+        # is unused, we also don't need it's decorators), we return the
+        # lineno of the first decorator, if there are any.
+        if first_node.decorator_list:
+            return first_node.decorator_list[0].lineno
+    return first_node.lineno
